backend/pyproject.toml
[tool.poetry]
name = "your-virtual-env"
version = "0.1.0"
description = "Your Description"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
pymongo = "^4.10.1"
python-dotenv = "^1.0.1"
fastapi = "^0.115.4"
uvicorn = "^0.32.0"
boto3 = "^1.35.70"
botocore = "^1.35.70"
voyageai = "^0.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
